Render histogram to report sheet
Note the perhaps "unorthodox" choice of placing an interactive component on the report sheet, as it's presented ostensibly as a print preview.
I've considered this drawback but I feel the potential information density outweigh the slight quirkiness.

Another conspicious choice in this commit is making a separate pass to collect the strictly zero values.
I went more in depth about that per email earlier. However we end up presenting the statistic I'll advocate for keeping this approach one way or the other.

I would like feedback on this specifically before merging. @gerdint

The French translation is strictly preliminary. Perhaps @bastienqueste could take a look at it.