Correcting unittest for percentile calibration
This testcase strangely seems to have been wrong for a long time.

The current dataset provides two values in "each end" of the histogram: bins 0 and 99 contains 3 measures each.

The previous test constructed a histogram that (presumedly unintentionally) excluded the actual max value, meaning only its low value got counted. The expected value 0.005 in the assertion was only "accidentally" right, but statistically wrong with respect to the dataset.
The expected value for the 95th percentile would actually be 1.0.

It may well be argued that some other dataset, preferably containing more values, should be used for the 'percentileNormalizer' test case.