SYM-513: Differential sankey charts in comparison
Introduce Sankey charts in the comparative report, similar to those rendered in the single report. The charts visualises the amount of increase and decrease of pressures unto specific nature values between the chosen calculation results.

Note that in the absence of pressures that increase or decrease with respect to any ecological value, the respective chart elements are excluded from the report.