Extend Hibernate dialect for Array types
Entity validation via Hibernate has been deficient since a couple of changes that were made to entities in commit 8a90d68 (part of release 1.10). This problem would cause deployment of the packaged app into Wildfly to fail, in case a database modeling the entities weren't already present.
After applying this commit the problem goes away.

We need to explicitly map any "array type" columns in the data entities to the numeric type code for arrays (2003) in a custom class, and replace the `hibernate.dialect` property in persistence.xml with the extended type.