SYM-606: Report unaccounted "overflow"
Positive intensity changes applied to pixels that are already at maximum value (100) are effectively 'thrown away' in the calculation step.
As an initial measure, we track which bands are affected by such 'overflow' and make the discrepancy apparent in the reports.

The core issue will be addressed more thoroughly in coming releases. It's yet to be decided how best to handle this problem.

Commit also contains some refactoring of the report sheets - introducing an "abstract" base class (Angular directive) for the report components. We also remove the obsoleted `bandMap` property from the Sankey chart rendering and replace it with the `selectMetaDisplayDictionary` selector.
Finally, the changeset also accomodates a slight bug fix in the mentioned Sankey chart component. Previously a misplaced method call within the chart template caused the graphic to render an arbitrary number of times "on top of itself", making the chart "links" more opaque than intended (ie, bluer).