SYM-581: Immediate batch calculation on areas
Commit includes changes to the data model: two mapped entity properties (`normalizationValue` and `areaMatrixMap`) are moved from `CalculationResult` into `ScenarioSnapshot`. Mapped property `latestCalculation` is removed from `ScenarioSnapshot`.

Two mapped properties are added to the `BatchCalculation` entity (`areasCalculation` and `areasOptions`). The property `scenarios` on the same entity are renamed to `entities` (so as to reflect that the specified entity ids may represent both `Scenarios` and `ScenarioAreas`)