Upgrade rxjs to v7
Upgraded rxjs in connection with the current task (SYM-633) to make benefit of `latestValueFrom($o)` in place of the deprecated .toPromise() pattern.

This change apparently also gets rid of some extraneous dependencies in the development build (can be observed in the package-lock.json diff), which makes things a little tidyer.