Removing unverified suggestion from README
It is not clear that this is actually possible using Wildfly, unless the frontend is to be bundled with the .war and served statically (which doesn't seem desirable, for various reasons).
At any rate, the method for serving the frontend as static content via the _Undertow_ component built into Wildfly as described by the linked article verifiably doesn't work, atleast not without some additional configuration - assuming that the UI is to be made available on the root path ("/").