SYM-699: JakartaEE upgrade to v10
Commit contains some code quality details not directly related to the upgrade, notably moving misplaced static methods into a couple of utility classes (`util\CalculationUtil.java`, `util\MetaDataUtil.java`).
This could quite probably be done more thoroughly, as there seems to be a number of instance methods that could (should) be made static, for performance and consistency.
The commit only deals with methods that were _already_ declared static on the injectable EJB classes to address the runtime warnings emitted on deployment, since it's inappropriate for beans.

There are a few tiny syntactical fixes as well.